ESP32 Initial Push
This is code for the ESP 32 that can take commands over MQTT parse them and plot the data on a OLED display. This is the code that I ended up using to complete the OLED HW. It still needs quite a few changes in order to meet the functionality the course project requires. Current changes needed
1. update pin assignments on the esp32
2. UART receive buffer needs to be implemented
3. Decoding the UART messages so they can be passed back to MQTT and the data can be used for the OLED